requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
coverage==5.3
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
iniconfig==1.1.1
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.1.1
PyYAML==5.3.1
regex==2020.10.28
requests==2.24.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
untokenize==0.1.1
urllib3==1.25.11
Werkzeug==1.0.1
